Add story for ReportBar, break ReportBar into separate component
