+++ src/components/Reports.jsx
@@ -1,7 +1,7 @@
 import axios from "axios"
 import { useEffect, useState } from "react"
 
-import '../styles/ReportBar.css';
+import ReportBar from 'ReportBar'
 
 export default function Reports(props) {
   const { projects } = props;
@@ -28,30 +28,10 @@ export default function Reports(props) {
     )
   })
 
-  const ReportBar = props => {
-    const {color, progress, goal} = props
-
-    const getPercentage = (num, denom) => {
-      const percent = (num/denom) * 100;
-      return Math.round(percent*1000) / 1000;
-    }
-
-    const style = {
-      backgroundColor: color
-    }
-
-    return (
-      <div className="reportBar" style={style}>
-        <div className="reportBar-inner">{getPercentage(progress, goal)}</div>
-      </div>
-    )
-  }
-
   return (
     <div className="reports container">
       <ul>
         {report}
-        
       </ul>
       <ReportBar
         progress={5}
